refactoring setupserver
